requirements..txt
streamlit
math
fractions
matplotlib.pyplot
numpy
plotly.express